# api/openapi/dbbridge.yaml
openapi: "3.1.0"

info:
  title: dbbridge
  description: Fault-tolerant SQL proxy that executes queries asynchronously and persists results for later retrieval.
  version: "1.0.0"

servers:
  - url: http://localhost:8080
    description: Local development

security:
  - bearerAuth: []

paths:
  /healthz:
    get:
      summary: Liveness probe
      operationId: healthz
      tags: [health]
      security: []
      responses:
        "200":
          description: Service is alive
          content:
            text/plain:
              schema:
                type: string
                example: OK

  /readyz:
    get:
      summary: Readiness probe
      operationId: readyz
      tags: [health]
      security: []
      responses:
        "200":
          description: Service is ready to accept traffic
          content:
            text/plain:
              schema:
                type: string
                example: READY
        "503":
          description: Service is not ready

  /metrics:
    get:
      summary: Prometheus metrics
      description: |
        Served on `server.admin_addr` when that is configured, together with
        `/v1/admin/*`; the metric labels enumerate every configured `db_id`, so
        the route needs the `admin` scope wherever it is mounted.
      operationId: metrics
      tags: [health]
      responses:
        "200":
          description: Prometheus text format metrics
          content:
            text/plain:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /v1/queries:
    post:
      summary: Start a query
      operationId: startQuery
      tags: [queries]
      parameters:
        - name: Idempotency-Key
          in: header
          description: Client-supplied key for deduplication. Repeated requests with the same key return the existing query.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartQueryRequest"
      responses:
        "202":
          description: Query accepted (async mode)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryRecord"
        "200":
          description: Query completed (sync mode)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryRecord"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Unknown database
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: Request body too large
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Too many concurrent queries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Instance is draining or the database is unreachable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /v1/queries/{id}:
    get:
      summary: Get query status
      operationId: getQueryStatus
      tags: [queries]
      parameters:
        - $ref: "#/components/parameters/queryId"
      responses:
        "200":
          description: Query record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryRecord"
        "404":
          description: Query not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /v1/queries/{id}:stop:
    post:
      summary: Stop a running query
      operationId: stopQuery
      tags: [queries]
      parameters:
        - $ref: "#/components/parameters/queryId"
      responses:
        "200":
          description: Query stopped
          content:
            application/json:
              schema:
                type: object
                properties:
                  query_id:
                    type: string
                  status:
                    type: string
                    example: STOPPED
        "404":
          description: Query not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /v1/queries/{id}/stats:
    get:
      summary: Get execution stats for a query
      operationId: getQueryStats
      tags: [queries]
      parameters:
        - $ref: "#/components/parameters/queryId"
      responses:
        "200":
          description: Query statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueryStats"
        "404":
          description: Query not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /v1/queries/{id}/result:
    get:
      summary: Download query result
      description: |
        Streams the query result. Supports HTTP Range requests for partial downloads.
        Responds with `206 Partial Content` and `Content-Range` header when a `Range` header is present.
      operationId: downloadResult
      tags: [queries]
      parameters:
        - $ref: "#/components/parameters/queryId"
        - name: offset
          in: query
          description: Byte offset to start reading from (ignored if Range header is present)
          schema:
            type: integer
            format: int64
        - name: limit
          in: query
          description: Maximum number of bytes to read (ignored if Range header is present)
          schema:
            type: integer
            format: int64
        - name: Range
          in: header
          description: "HTTP byte range, e.g. `bytes=0-1023`"
          schema:
            type: string
      responses:
        "200":
          description: Full result stream
          content:
            application/x-jsonlines:
              schema:
                type: string
                format: binary
            text/csv:
              schema:
                type: string
                format: binary
            application/octet-stream:
              schema:
                type: string
                format: binary
        "206":
          description: Partial content (Range request)
          headers:
            Content-Range:
              description: "Byte range returned, e.g. `bytes 0-1023/4096`"
              schema:
                type: string
          content:
            application/x-jsonlines:
              schema:
                type: string
                format: binary
            text/csv:
              schema:
                type: string
                format: binary
            application/octet-stream:
              schema:
                type: string
                format: binary
        "400":
          description: The query has not succeeded, so no result exists yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Query or result not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "416":
          description: |
            Range not satisfiable. The `Content-Range` header carries
            `bytes */<total>`.
          headers:
            Content-Range:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /v1/databases:
    get:
      summary: List configured databases
      operationId: listDatabases
      tags: [databases]
      responses:
        "200":
          description: List of database infos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DatabaseInfo"
        "500":
          description: Internal error
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /v1/ws:
    get:
      summary: WebSocket endpoint for query event subscriptions
      description: |
        Upgrade to WebSocket. After upgrade, send `{"action":"watch","query_id":"<id>"}` to subscribe
        and `{"action":"unwatch","query_id":"<id>"}` to drop a subscription.
        The server streams `{"query_id":"...","state":"...","stats":{...},"error":null}` as state changes.

        The current state is delivered immediately on subscribe, and the subscription is closed once
        the query reaches a terminal state. A connection may hold at most 32 subscriptions; further
        `watch` messages are answered with an error message. Cross-origin connections are refused
        unless the origin is listed in `server.ws_allowed_origins`.

        A browser cannot set the `Authorization` header on a handshake, so the credential may also
        be offered as a subprotocol pair: `new WebSocket(url, ["dbbridge.bearer", token])`. The
        server selects and echoes only `dbbridge.bearer`, never the token.
      operationId: wsSubscribe
      tags: [queries]
      responses:
        "101":
          description: Switching Protocols
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /v1/admin/reload:
    post:
      summary: Hot-reload configuration
      operationId: reloadConfig
      tags: [admin]
      responses:
        "200":
          description: Config reloaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  message:
                    type: string
                  report:
                    $ref: "#/components/schemas/ReloadReport"
        "500":
          description: Reload failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  /v1/admin/can-stop:
    get:
      summary: Check if instance can be stopped safely
      description: Returns `can_be_stopped=true` only when there are zero in-flight owned queries (invariant I5).
      operationId: canIBeStopped
      tags: [admin]
      responses:
        "200":
          description: Stop readiness status
          content:
            application/json:
              schema:
                type: object
                properties:
                  can_be_stopped:
                    type: boolean
                  in_flight:
                    type: integer
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

components:
  responses:
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: The token lacks the scope this operation requires
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        Static bearer token from `auth.tokens`. Each token carries a subject and
        a set of scopes: `read` (status, stats, download, list, watch), `write`
        (submit, stop) and `admin` (reload, can-stop, `/metrics`). `write`
        implies `read`, so a token that submits a query can poll it and fetch
        its result; `admin` implies both and can act across subjects, while
        every other token only reaches the queries its own subject submitted.
        Endpoints answer `401` without a valid token and `403` when the token
        lacks the scope. The health probes are exempt.

        Revoking a token requires a restart: `auth` is reported under
        `report.ignored` by `POST /v1/admin/reload` rather than applied.

        Turning authentication on binds new records to a subject but leaves
        older ones with an empty one, and those are reachable only by `admin`.
        Roll it out on an idle instance, or expect in-flight and retained
        queries submitted before the switch to answer `404` to their owners
        until `result_ttl` expires.

  parameters:
    queryId:
      name: id
      in: path
      required: true
      description: Query ID
      schema:
        type: string

  schemas:
    StartQueryRequest:
      type: object
      required: [database_id, sql]
      properties:
        database_id:
          type: string
          description: ID of the target database as configured in dbbridge.yaml
        sql:
          type: string
          description: SQL statement to execute
        options:
          type: object
          properties:
            timeout_ms:
              type: integer
              format: int64
              description: Query timeout in milliseconds; 0 means no timeout
            mode:
              type: string
              enum: [async, sync]
              default: async
              description: async returns immediately with a query_id; sync waits for completion
            result_ttl_seconds:
              type: integer
              format: int64
              description: Seconds to retain the result; 0 uses server default
            result_format:
              type: string
              enum: [jsonl, csv]
              default: jsonl
              description: |
                Validated against this list before the query runs. `parquet` is
                declared in the spec but not implemented, so it is rejected.
            storage_backend:
              type: string
              enum: [fs, s3, clickhouse]
              description: Override the default storage backend for this query

    QueryRecord:
      type: object
      properties:
        id:
          type: string
        database_id:
          type: string
        sql:
          type: string
        state:
          $ref: "#/components/schemas/QueryState"
        owner_instance_id:
          type: string
        options:
          $ref: "#/components/schemas/QueryOptions"
        stats:
          $ref: "#/components/schemas/QueryStats"
        result:
          $ref: "#/components/schemas/ResultRef"
        error:
          $ref: "#/components/schemas/QueryError"
        created_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        finished_at:
          type: string
          format: date-time
        lease_deadline:
          type: string
          format: date-time
          description: Derived from the owner's lease key, absent once the query is terminal
        idempotency_key:
          type: string

    Error:
      type: object
      description: |
        The single error shape of the API. `error` carries the category and,
        for input errors, the reason; internal detail stays in the server log,
        keyed by `request_id`.
      required: [error]
      properties:
        error:
          type: string
        request_id:
          type: string

    ReloadReport:
      type: object
      properties:
        added:
          type: array
          items:
            type: string
        removed:
          type: array
          items:
            type: string
        updated:
          type: array
          items:
            type: string
        ignored:
          type: array
          description: Sections that changed but require a restart to take effect
          items:
            type: string
        failures:
          type: array
          description: Databases whose pool could not be opened
          items:
            type: string

    QueryState:
      type: string
      enum: [PENDING, RUNNING, SUCCEEDED, FAILED, CANCELED, EXPIRED]

    QueryOptions:
      type: object
      properties:
        timeout_ms:
          type: integer
          format: int64
        mode:
          type: string
        result_ttl_seconds:
          type: integer
          format: int64
        result_format:
          type: string
        storage_backend:
          type: string
        idempotency_key:
          type: string

    QueryStats:
      type: object
      properties:
        rows_read:
          type: integer
          format: int64
        bytes_written:
          type: integer
          format: int64
        db_exec_duration_ms:
          type: integer
          format: int64
        storage_write_duration_ms:
          type: integer
          format: int64
        total_duration_ms:
          type: integer
          format: int64
        retries:
          type: integer

    ResultRef:
      type: object
      properties:
        backend:
          type: string
        locator:
          type: string
        size_bytes:
          type: integer
          format: int64
        row_count:
          type: integer
          format: int64
        format:
          type: string
        checksum:
          type: string

    QueryError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        retryable:
          type: boolean

    DatabaseInfo:
      type: object
      properties:
        id:
          type: string
        engine:
          type: string
          enum: [postgres, mysql, clickhouse, oracle]
        display_name:
          type: string
        healthy:
          type: boolean